:arrow_up: Upgrade iced 0.12 → 0.14 to fix scrollbar release
iced 0.12's scrollable processed children before its own grabbed-state
release handler, so a child that captured `ButtonReleased(Left)` (e.g.
`text_editor`, which captures it unconditionally regardless of cursor
position) blocked the scrollbar from ever clearing `y_scroller_grabbed_at`.
iced 0.13 fixed this by always clearing the grabbed state on left release
(scrollable.rs:734-747).

Migration touchpoints:
- `Application` trait → `iced::application(boot, update, view).run()` builder
- `Command` → `Task`; `Command::perform/none/batch` → `Task::*` equivalents
- `subscription::channel` → `Subscription::run(|| stream::channel(...))`
- `iced::theme::Container::Box` → `container::bordered_box` style closure
- `Row::align_items` → `align_y`; `Column::align_items` → `align_x`
- `Space::with_width(L)` → `Space::new().width(L)` (helper renamed)
- `Rule::horizontal(n)` → `rule::horizontal(n)` (helper module)
- `checkbox(label, checked)` → `checkbox(checked).label(label)`
- `iced::window::Id::MAIN` removed → `window::latest().and_then(...)`
- `window::change_mode` → `window::set_mode`
- Padding 4-tuple `[t, r, b, l]` no longer impls `Into<Padding>` →
  `Padding::default().right(...)`
- `Pixels: From<u16>` removed (only `u32`/`f32`) → theme spacing
  constants switched to `f32`
- `text()` wrapper now takes `IntoFragment<'a>` instead of `Cow<'a, str>`

+++ src/theme.rs
@@ -11,10 +11,10 @@ pub fn celeste_theme() -> Theme {
 }
 
 /// Standard outer padding for pages.
-pub const PAGE_PADDING: u16 = 16;
+pub const PAGE_PADDING: f32 = 16.0;
 
 /// Spacing between stacked sections within a page.
-pub const SECTION_SPACING: u16 = 12;
+pub const SECTION_SPACING: f32 = 12.0;
 
 /// Spacing between items inside a row.
-pub const ROW_SPACING: u16 = 8;
+pub const ROW_SPACING: f32 = 8.0;
